+++ Inf Grafica/Sources/vector.cpp
@@ -10,10 +10,10 @@
 
 bool Vector::esVector() { return true; } 
 
-float Vector::modulo() { return sqrt(pow(valores[0], 2) + pow(valores[1], 2) + pow(valores[2], 2)); }
+double Vector::modulo() { return sqrt(pow(valores[0], 2) + pow(valores[1], 2) + pow(valores[2], 2)); }
 
 void Vector::normalizar(){
-	float module = modulo();
+	double module = modulo();
 	set_values(valores[0]/module, valores[1]/module, valores[2]/module);
 }
